+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Fixed
+
+- The marker icon passed to the Leaflet constructor is now only replaced with an empty div when there is
+  additional content within the `<LMarker>` component, _and_ at least some of that content is something
+  other than a tooltip, resolving
+  [#269 Marker with tooltip no longer visible with v0.8.2](https://github.com/vue-leaflet/vue-leaflet/issues/269).
+
+
 ## [0.8.2] - 2023-02-05
 
 ### Fixed